META-244: Fixed test AnalyticsDbCursor.reads_all_matched_data
META-244: Implemented analytics db cursor closure
VMS-12734: encoder models added
VMS-13743: Implement system health notification event for metadata storage
VMS-13842: CloudStreamSocket::pollable() returned nullptr after CloudStreamSocket::shutdown()
VMS-13924. Server has crashed on Windows: c0000005 (Access violation)
VMS-13935: Windows mediaserver is constantly crashing (build 28936)
VMS-13940: Server has crashed on Windows: c0000005 (Access violation
WEB-1004: When user tries to login with non-existing username there is wrong error message shown
WEB-999: The 'Playing position' is displayed incorrectly.

<==== Mercurial data ====>
Branch: vms_4.0_gui

+++ vms/server/nx_vms_server/src/plugins/resource/d-link/dlink_resource_searcher.cpp
@@ -86,7 +86,7 @@ QnResourceList QnPlDlinkResourceSearcher::findResources()
             continue;
 
         // Sending broadcast.
-        constexpr qint16 kPort = 62976;
+        constexpr qint16 kPort = qint16(62976);
         for (int r = 0; r < CL_BROAD_CAST_RETRY; ++r)
         {
             sock->sendTo(barequest.data(), barequest.size(), nx::network::BROADCAST_ADDRESS, kPort);
